# src/main/java/com/perezper/authserver/dto/AuthResponse.java
package com.perezper.authserver.dto;

public class AuthResponse {
    private String accessToken;
    private String refreshToken;
    private long expiresIn;

    public AuthResponse() {}
    public AuthResponse(String accessToken, String refreshToken, long expiresIn) {
        this.accessToken = accessToken; this.refreshToken = refreshToken; this.expiresIn = expiresIn;
    }
    public String getAccessToken() { return accessToken; }
    public void setAccessToken(String accessToken) { this.accessToken = accessToken; }
    public String getRefreshToken() { return refreshToken; }
    public void setRefreshToken(String refreshToken) { this.refreshToken = refreshToken; }
    public long getExpiresIn() { return expiresIn; }
    public void setExpiresIn(long expiresIn) { this.expiresIn = expiresIn; }
}