# src/main/java/com/perezper/authserver/dto/ErrorResponse.java
package com.perezper.authserver.dto;

public class ErrorResponse {
    private String error;

    public ErrorResponse() {}
    public ErrorResponse(String error) { this.error = error; }
    public String getError() { return error; }
    public void setError(String error) { this.error = error; }
}