# src/main/java/com/perezper/authserver/dto/RegisterRequest.java
package com.perezper.authserver.dto;

public class RegisterRequest {
    private String username;
    private String password;
    private String email;

    public RegisterRequest() {}
    public String getUsername() { return username; }
    public void setUsername(String username) { this.username = username; }
    public String getPassword() { return password; }
    public void setPassword(String password) { this.password = password; }
    public String getEmail() { return email; }
    public void setEmail(String email) { this.email = email; }
}